Modify transfer network and forward pass to make dimensions match
After this change, each batch results in a single scalar output from the network,
to guarantee accurate broadcasting it is done manually. The x tensor, which started
out as (batch size x block_size x n_embd) is flattened along the second and third
dimensions, and then the output from the transfer sequential block  is squeezed down
to just the batch dimension before it is used to compute the loss.

We are now able to use a larger batch size, because of the padding changes added earlier.
However, overall this change slows down the training job a lot.